+++ src/ACT.Application/Services/Interfaces/IJwtService.cs
@@ -1,0 +1,9 @@
+using ACT.Domain.Entities;
+
+namespace ACT.Application.Services.Interfaces;
+
+public interface IJwtService
+{
+    string GenerateToken(User user);
+}
+